Extract views router module
